requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-cloud-speech
google-cloud-texttospeech
google_auth_oauthlib
google-crc32c
google-genai
google-generativeai
googleapis-common-protos
python-multipart